new functionality
Updated beam correction code. Code now does beam correction automatically based comparison of the current phase center with the beam correction direction stored in the ms by DP3. --beamcor='auto' is the default so that the user does not have to worry about it anymore ('yes', or 'no' can also be used to overwrite the 'auto' setting'). Option no-beamcor removed. Updated ms POLARIZATION table info based on circular and linear conversions so that the Stokes Q, U, and V naming is always correct.